ddti-sequencer/requirements_synth.txt
# Core MIDI handling
mido==1.3.2
python-rtmidi==1.5.8

# Audio synthesis dependencies
numpy>=1.24.0
sounddevice>=0.4.6

# Optional: FluidSynth (for realistic bass samples)
# pyfluidsynth>=1.3.2  # Uncomment if using FluidSynth version


